+++ apps/generic_apps/pltt_app/PLTT_privacy_tracker_target_app.cpp
@@ -20,7 +20,7 @@
 
 typedef wiselib::OSMODEL Os;
 typedef Os::TxRadio Radio;
-typedef uint8 CoordinatesNumber;
+typedef uint16 CoordinatesNumber;
 typedef uint8 IntensityNumber;
 typedef uint8 TimesNumber;
 typedef uint8 SecondsNumber;
